Reverted to commit ca62a2d66.
Let dzil build the files of the distribution.

Commit 94d5a6e31 added files that should be generated by "dzil build".
This commit deletes files added, and restores files which was modified
to the version in commit ca62a2d66.

A small problem that this commit introduces is how to include the last
commit c2ca1e9c6 ?
This commit was added after 94d5a6e31 and modifies the README file.
However the README file is currently generated by "dzil build" by plugin
[Readme]. Some options to handle this is a) to remove the [Readme] plugin,
such that the Readme file is not genereated. But this has a disadvantage
that the README must be updated manually if the Copyright year or
version of the module changes, b) write a custom plugin that modifies
the generated README, c) use another plugin like [ReadmeAnyFromPod], d)
put the added information in some other file, e) abandon the added
infomration all together.